+++ pysat/instruments/methods/testing.py
@@ -1,34 +1,106 @@
 import datetime as dt
 import numpy as np
 import os
+import pandas as pds
 import warnings
 
-import pandas as pds
+import pysat
 
-from pysat.utils import NetworkLock
-from pysat import here
+logger = pysat.logger
 
 ackn_str = ' '.join(("Test instruments provided through the pysat project.",
                      "https://www.github.com/pysat/pysat"))
 
 # Load up citation information
-with NetworkLock(os.path.join(here, 'citation.txt'), 'r') as locked_file:
+with pysat.utils.NetworkLock(os.path.join(pysat.here, 'citation.txt'), 'r') as \
+        locked_file:
     refs = locked_file.read()
 
 
+def init(self, test_init_kwarg=None):
+    """Initializes the Instrument object with instrument specific values.
+
+    Runs once upon instantiation.
+
+    Shifts time index of files by 5-minutes if mangle_file_dates
+    set to True at pysat.Instrument instantiation.
+
+    Creates a file list for a given range if the file_date_range
+    keyword is set at instantiation.
+
+    Parameters
+    ----------
+    self : pysat.Instrument
+        This object
+    test_init_kwarg : any or NoneType
+        Testing keyword (default=None)
+
+    """
+
+    logger.info(ackn_str)
+    self.acknowledgements = ackn_str
+    self.references = refs
+
+    # Assign parameters for testing purposes
+    self.new_thing = True
+    self.test_init_kwarg = test_init_kwarg
+
+    return
+
+
+def clean(self, test_clean_kwarg=None):
+    """Cleaning function
+
+    Parameters
+    ----------
+    self : pysat.Instrument
+        This object
+    test_clean_kwarg : any or NoneType
+        Testing keyword (default=None)
+
+    """
+
+    self.test_clean_kwarg = test_clean_kwarg
+
+    return
+
+
+# Optional method
+def preprocess(self, test_preprocess_kwarg=None):
+    """Customization method that performs standard preprocessing.
+
+    This routine is automatically applied to the Instrument object
+    on every load by the pysat nanokernel (first in queue). Object
+    modified in place.
+
+    Parameters
+    ----------
+    self : pysat.Instrument
+        This object
+    test_preprocess_kwarg : any or NoneType
+        Testing keyword (default=None)
+
+    """
+
+    self.test_preprocess_kwarg = test_preprocess_kwarg
+
+    return
+
+
 def list_files(tag=None, inst_id=None, data_path=None, format_str=None,
-               file_date_range=None, test_dates=None):
+               file_date_range=None, test_dates=None, mangle_file_dates=False,
+               test_list_files_kwarg=None):
     """Produce a fake list of files spanning three years
 
     Parameters
     ----------
-    tag : str
+    tag : str or NoneType
         pysat instrument tag (default=None)
-    inst_id : str
+    inst_id : str or NoneType
         pysat satellite ID tag (default=None)
-    data_path : str
+    data_path : str or NoneType
         pysat data path (default=None)
-    format_str : str
+    format_str : str or NoneType
         file format string (default=None)
     file_date_range : pds.date_range
         File date range. The default mode generates a list of 3 years of daily
@@ -36,14 +108,22 @@ def list_files(tag=None, inst_id=None, data_path=None, format_str=None,
         through below.  Otherwise, accepts a range of files specified by the
         user.
         (default=None)
-    test_dates : dt.datetime
+    test_dates : dt.datetime or NoneType
         Pass the _test_date object through from the test instrument files
+    mangle_file_dates : bool
+        If True, file dates are shifted by 5 minutes. (default=False)
+    test_list_files_kwarg : any or NoneType
+        Testing keyword (default=None)
 
     Returns
     -------
     Series of filenames indexed by file time
 
     """
+
+    # Support keyword testing
+    logger.info(''.join(('test_list_files_kwarg = ',
+                         str(test_list_files_kwarg))))
 
     if data_path is None:
         data_path = ''
@@ -57,6 +137,10 @@ def list_files(tag=None, inst_id=None, data_path=None, format_str=None,
 
     index = file_date_range
 
+    # Mess with file dates if kwarg option set
+    if mangle_file_dates:
+        index = index + dt.timedelta(minutes=5)
+
     # Create the list of fake filenames
     names = [data_path + date.strftime('%Y-%m-%d') + '.nofile'
              for date in index]
@@ -66,19 +150,20 @@ def list_files(tag=None, inst_id=None, data_path=None, format_str=None,
 
 def list_remote_files(tag=None, inst_id=None, data_path=None, format_str=None,
                       start=None, stop=None, test_dates=None, user=None,
-                      password=None):
+                      password=None, mangle_file_dates=False,
+                      test_list_remote_kwarg=None):
     """Produce a fake list of files spanning three years and one month to
     simulate new data files on a remote server
 
     Parameters
     ----------
-    tag : str
+    tag : str or NoneType
         pysat instrument tag (default=None)
-    inst_id : str
+    inst_id : str or NoneType
         pysat satellite ID tag (default=None)
-    data_path : str
+    data_path : str or NoneType
         pysat data path (default=None)
-    format_str : str
+    format_str : str or NoneType
         file format string (default=None)
     start : dt.datetime or NoneType
         Starting time for file list. A None value will start 1 year before
@@ -88,14 +173,18 @@ def list_remote_files(tag=None, inst_id=None, data_path=None, format_str=None,
         Ending time for the file list.  A None value will stop 2 years 1 month
         after test_date
         (default=None)
-    test_dates : dt.datetime
+    test_dates : dt.datetime or NoneType
         Pass the _test_date object through from the test instrument files
-    user : string
+    user : string or NoneType
         User string input used for download. Provided by user and passed via
         pysat. If an account is required for dowloads this routine here must
         error if user not supplied. (default=None)
-    password : string
+    password : string or NoneType
         Password for data download. (default=None)
+    mangle_file_dates : bool
+        If True, file dates are shifted by 5 minutes. (default=False)
+    test_list_remote_kwarg : any or NoneType
+        Testing keyword (default=None)
 
     Returns
     -------
@@ -103,6 +192,10 @@ def list_remote_files(tag=None, inst_id=None, data_path=None, format_str=None,
         Filenames indexed by file time, see list_files for more info
 
     """
+
+    # Support keyword testing
+    logger.info(''.join(('test_list_remote_kwarg = ',
+                         str(test_list_remote_kwarg))))
 
     # Determine the appropriate date range for the fake files
     if start is None:
@@ -116,11 +209,12 @@ def list_remote_files(tag=None, inst_id=None, data_path=None, format_str=None,
 
     return list_files(tag=tag, inst_id=inst_id, data_path=data_path,
                       format_str=format_str, file_date_range=file_date_range,
+                      mangle_file_dates=mangle_file_dates,
                       test_dates=test_dates)
 
 
 def download(date_array, tag, inst_id, data_path=None, user=None,
-             password=None):
+             password=None, test_download_kwarg=None):
     """Simple pass function for pysat compatibility for test instruments.
 
     This routine is invoked by pysat and is not intended for direct use by the
@@ -135,20 +229,18 @@ def download(date_array, tag, inst_id, data_path=None, user=None,
         Tag identifier used for particular dataset. This input is provided by
         pysat. (default='')
     inst_id : string
-        Satellite ID string identifier used for particular dataset. This input
+        Instrument ID string identifier used for particular dataset. This input
         is provided by pysat. (default='')
-    data_path : string
+    data_path : string or NoneType
         Path to directory to download data to. (default=None)
-    user : string
+    user : string or NoneType
         User string input used for download. Provided by user and passed via
-        pysat. If an account is required for dowloads this routine here must
+        pysat. If an account is required for downloads this routine here must
         error if user not supplied. (default=None)
-    password : string
+    password : string or NoneType
         Password for data download. (default=None)
-    **kwargs : dict
-        Additional keywords supplied by user when invoking the download
-        routine attached to a pysat.Instrument object are passed to this
-        routine via kwargs.
+    test_download_kwarg : any or NoneType
+        Testing keyword (default=None)
 
     Raises
     ------
@@ -160,6 +252,9 @@ def download(date_array, tag, inst_id, data_path=None, user=None,
     When no download support will be provided
 
     """
+
+    # Support keyword testing
+    logger.info(''.join(('test_download_kwarg = ', str(test_download_kwarg))))
 
     if tag == 'no_download':
         warnings.warn('This simulates an instrument without download support')